+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog for package rosros
 * fix api.dict_to_message() erroring on temporal types in dict
 * fix util.get_arity() counting some parameters twice if counting both positional and keyword arguments
 * fix api.time_message() not using clock_type parameter in ROS2
+* fix patching floordiv to ROS2 duration class
 
 0.2.0 (2023-02-28)
 -------------------